+++ automerge-backend/src/op_handle.rs
@@ -4,8 +4,8 @@ use std::hash::{Hash, Hasher};
 use std::ops::Deref;
 use std::rc::Rc;
 
-use crate::protocol::{Change, Operation, UndoOperation};
-use automerge_protocol::{OpID, ObjectID, Key, Value, OpType};
+use crate::protocol::{Change, UndoOperation};
+use automerge_protocol::{OpID, ObjectID, Key, Value, OpType, Operation};
 
 #[derive(Clone)]
 pub(crate) struct OpHandle {
